+++ src/components/index.js
@@ -7,3 +7,4 @@ export { default as NrqlEditor } from './nrql-editor';
 export { default as StatusIcon } from './status-icon';
 export { default as StatusIconsLayout } from './status-icons-layout';
 export { default as ProgressBar } from './progress-bar';
+export { default as FilterBar } from './filter-bar';