+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- FilterBar component by [@amit-y](https://github.com/amit-y)
+
 ## [1.15.0] - 2023-05-18
 
 ### Added